完善基础战斗框架

+++ Assets/Scripts/Battle/Manager/DamageManager.cs
@@ -7,8 +7,8 @@ namespace GameCore {
 
     }
 
-    public async UniTask Damage(Unit source, Unit target, float damageValue) {
-      int realDamageValue = target.AddAttrib(AttribType.HP, -(int)damageValue);
+    public async UniTask Damage(Unit source, Unit target, int damageValue) {
+      int realDamageValue = target.AddAttrib(AttribType.HP, -damageValue);
       // Test
       Debug.Log($"{source.RuntimeId}:{source.Name} 对 {target.RuntimeId}:{target.Name} 造成{-realDamageValue}点伤害");
       if (target.GetAttrib(AttribType.HP).Value <= 0) {
